+++ app/views/pages/index_one_page_elevator.html.erb
@@ -1312,8 +1312,7 @@
 								<strong>Sorry!</strong> You need to complete all mandatory (*) fields!
 							</div><!-- /Alert Mandatory -->
 
-
-							<form action="/leads" method="post">
+							<form action="/leads" method="post" enctype="multipart/form-data">
 								<fieldset>
 									<input type="hidden" name="action" value="contact_send" />
 
@@ -1366,14 +1365,12 @@
 											</div>
 									</div>
 									<div class="row">
-											<div class="col-md-12">
-												<label for="contact:attachment">File Attachment</label>
-
-												<!-- custom file upload -->
-												<input class="custom-file-upload" type="file" id="file" name="Attachment" id="contact:attachment" data-btn-text="Select a File" />
-												<small class="text-muted block">Max file size: 10Mb (zip/pdf/jpg/png)</small>
-
-											</div>
+										<div class="col-md-12">
+											<label for="attachment">File Attachment</label>
+											<!-- custom file upload -->
+											<input class="custom-file-upload" type="file" id="file" name="attachment" id="attachment" data-btn-text="Select a File" />
+											<small class="text-muted block">Max file size: 10Mb (zip/pdf/jpg/png)</small>
+										</div>
 									</div>
 
 								</fieldset>
