+++ db/migrate/20200327170925_add_file_name_to_leads.rb
@@ -1,0 +1,5 @@
+class AddFileNameToLeads < ActiveRecord::Migration[5.2]
+  def change
+    add_column :leads, :file_name, :string
+  end
+end
