+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_03_26_155640) do
+ActiveRecord::Schema.define(version: 2020_03_27_170925) do
 
   create_table "active_storage_attachments", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
     t.string "name", null: false
@@ -229,6 +229,7 @@ ActiveRecord::Schema.define(version: 2020_03_26_155640) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.binary "attachment"
+    t.string "file_name"
   end
 
   create_table "quotes", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
